backend/common/serializers.py
from recipes.models import Recipe
from rest_framework import serializers


class RecipeMiniSerializer(serializers.ModelSerializer):

    class Meta:
        model = Recipe
        fields = ('id', 'name', 'image', 'cooking_time')
