backend/requirements.txt
Django==3.2.16
django-filter==22.1
djangorestframework==3.12.4
djoser==2.1.0
django-cors-headers==3.13.0
psycopg2-binary==2.9.3
webcolors==1.11.1
Pillow==9.0.0
django-colorfield
drf-extra-fields